split out tracker implementations
